Preserve strides of custom Triton kenrel args
Summary: Currently, we [`clone`](https://github.com/pytorch/pytorch/blob/19207b9183baced683d675c341b7da051cd16440/torch/_inductor/lowering.py#L5273) every `TensorBox` argument of custom Triton kernels during lowering to the Inductor IR, during which the stride information is lost. This is problematic for the common case when the strides of a `torch.Tensor` argument are passed as scalars to a custom Triton kenrel alongside the tensor itself (due to the underlying Triton code interpreting the tensors as row pointers, so the contained stride semantics of `torch.Tensor` is lost).

In this PR, we add another version of `clone`---`clone_preserve_reinterpret_view`---which does carries over the `ir.ReinterpretVew` layers from the source `TensorBoxed` to the cloned one. The rationale of adding a new function (and switching to it in the `triton_kernel_wrap` only for now) is keeping the semantics of the existing `clone` untouched, as it is a lowering of `torch.clone` (albeit incomplete, as the `memory_format` is currently ignored): changing the exiting `clone` would change the semantics which is not necessarily desirable in general. Open to suggestions, though.

Test Plan:

```
$ python test/dynamo/test_functions.py -k test_triton_kernel_strided_input
...
----------------------------------------------------------------------
Ran 1 test in 5.568s

OK

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 27c5532d2195fb87ecb8ee3dd0b2b0882a7f99db
Pull Request resolved: https://github.com/pytorch/pytorch/pull/116219